Update beforeMbLayerId to addLayer
